+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 -   The new `sea_ice_index()` function downloads monthly and daily sea ice extent.
+-   `use_cache` is now `TRUE` by default.
 
 # icecdr 1.0.0